+++ packages/interledger/src/accounts/errors.ts
@@ -1,19 +1,5 @@
 import { Asset } from './types'
 
-export class BalanceTransferError extends Error {
-  constructor(public code: number) {
-    super()
-    this.name = 'TransferError'
-  }
-}
-
-export class UnknownBalanceError extends Error {
-  constructor(accountId: string) {
-    super('Balance not found. accountId=' + accountId)
-    this.name = 'UnknownBalanceError'
-  }
-}
-
 export class UnknownLiquidityAccountError extends Error {
   constructor(asset: Asset) {
     super(
