+++ packages/interledger/src/index.ts
@@ -9,6 +9,7 @@ import { createRatesService, Rafiki } from './connector/core'
 import { Config } from './config'
 import { AccountsService } from './accounts/service'
 import { createBalanceService } from './balance/service'
+import { createCreditService } from './credit/service'
 import { Logger } from './logger/service'
 
 const logger = Logger
@@ -66,13 +67,18 @@ export const start = async (): Promise<void> => {
 
   const accountsService = new AccountsService(balanceService, Config, logger)
 
+  const creditService = createCreditService({
+    logger,
+    balanceService
+  })
+
   const ratesService = createRatesService({
     pricesUrl,
     pricesLifetime,
     logger
   })
 
-  adminApi = await createAdminApi({ accountsService, logger })
+  adminApi = await createAdminApi({ accountsService, creditService, logger })
 
   connectorApp = await createConnectorService({
     redis,
