+++ packages/interledger/src/shared/errors.ts
@@ -1,0 +1,13 @@
+export class BalanceTransferError extends Error {
+  constructor(public code: number) {
+    super()
+    this.name = 'TransferError'
+  }
+}
+
+export class UnknownBalanceError extends Error {
+  constructor(accountId: string) {
+    super('Balance not found. accountId=' + accountId)
+    this.name = 'UnknownBalanceError'
+  }
+}
